First pass at showing program card in admin program image upload
