+++ server/app/views/admin/programs/ProgramImageView.java
@@ -3,31 +3,42 @@ package views.admin.programs;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static j2html.TagCreator.div;
 import static j2html.TagCreator.form;
-import static j2html.TagCreator.img;
+import static j2html.TagCreator.h2;
 
+import auth.CiviFormProfile;
+import auth.ProfileUtils;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.Inject;
 import com.typesafe.config.Config;
 import controllers.admin.routes;
+import controllers.applicant.ApplicantRoutes;
 import forms.admin.ProgramImageDescriptionForm;
 import j2html.tags.specialized.DivTag;
 import j2html.tags.specialized.FormTag;
-import j2html.tags.specialized.ImgTag;
 import j2html.tags.specialized.InputTag;
+import j2html.tags.specialized.LiTag;
+import java.time.ZoneId;
 import java.util.Optional;
 import play.data.Form;
 import play.data.FormFactory;
+import play.i18n.Messages;
+import play.i18n.MessagesApi;
 import play.mvc.Http;
 import play.twirl.api.Content;
 import services.LocalizedStrings;
+import services.MessageKey;
+import services.applicant.ApplicantPersonalInfo;
+import services.applicant.ApplicantService;
 import services.cloud.PublicFileNameFormatter;
 import services.cloud.PublicStorageClient;
 import services.cloud.StorageUploadRequest;
 import services.program.ProgramDefinition;
+import services.settings.SettingsManifest;
 import views.BaseHtmlView;
 import views.HtmlBundle;
 import views.admin.AdminLayout;
 import views.admin.AdminLayoutFactory;
+import views.applicant.ProgramCardView;
 import views.components.ButtonStyles;
 import views.components.FieldWithLabel;
 import views.components.ToastMessage;
@@ -42,22 +53,38 @@ public final class ProgramImageView extends BaseHtmlView {
 
   private final AdminLayout layout;
   private final String baseUrl;
+  private final ApplicantRoutes applicantRoutes;
   private final FormFactory formFactory;
   private final FileUploadViewStrategy fileUploadViewStrategy;
+  private final MessagesApi messagesApi;
+
+  private final ProfileUtils profileUtils;
   private final PublicStorageClient publicStorageClient;
+  private final SettingsManifest settingsManifest;
+  private final ZoneId zoneId;
 
   @Inject
   public ProgramImageView(
       AdminLayoutFactory layoutFactory,
+      ApplicantRoutes applicantRoutes,
       Config config,
       FormFactory formFactory,
       FileUploadViewStrategy fileUploadViewStrategy,
-      PublicStorageClient publicStorageClient) {
+      MessagesApi messagesApi,
+      ProfileUtils profileUtils,
+      PublicStorageClient publicStorageClient,
+      SettingsManifest settingsManifest,
+      ZoneId zoneId) {
     this.layout = checkNotNull(layoutFactory).getLayout(AdminLayout.NavPage.PROGRAMS);
     this.baseUrl = checkNotNull(config).getString("base_url");
+    this.applicantRoutes = applicantRoutes;
     this.formFactory = checkNotNull(formFactory);
     this.fileUploadViewStrategy = checkNotNull(fileUploadViewStrategy);
     this.publicStorageClient = checkNotNull(publicStorageClient);
+    this.messagesApi = checkNotNull(messagesApi);
+    this.profileUtils = checkNotNull(profileUtils);
+    this.settingsManifest = checkNotNull(settingsManifest);
+    this.zoneId = zoneId;
   }
 
   /**
@@ -68,15 +95,21 @@ public final class ProgramImageView extends BaseHtmlView {
     String title =
         String.format(
             "Manage program image for %s", programDefinition.localizedName().getDefault());
-    HtmlBundle htmlBundle =
-        layout
-            .getBundle(request)
-            .setTitle(title)
-            .addMainContent(
-                renderHeader(title, "my-10", "mx-10"),
-                createImageDescriptionForm(request, programDefinition),
-                createImageUploadForm(programDefinition),
-                renderCurrentImage(programDefinition));
+
+    DivTag mainContent =
+        div()
+            .withClasses("my-10", "mx-20")
+            .with(renderHeader(title))
+            .with(createImageDescriptionForm(request, programDefinition));
+
+    DivTag imageUploadAndCurrentCardContainer =
+        div()
+            .withClasses("grid", "grid-cols-2", "w-full")
+            .with(createImageUploadForm(programDefinition))
+            .with(renderCurrentProgramCard(request, programDefinition));
+    mainContent.with(imageUploadAndCurrentCardContainer);
+
+    HtmlBundle htmlBundle = layout.getBundle(request).setTitle(title).addMainContent(mainContent);
 
     // TODO(#5676): This toast code is re-implemented across multiple controllers. Can we write a
     // helper method for it?
@@ -139,7 +172,6 @@ public final class ProgramImageView extends BaseHtmlView {
             // for more context.
             .with(fileInputElement);
 
-    // TODO(#5676): Replace with final UX once we have it.
     return div()
         .with(fullForm)
         .with(
@@ -159,18 +191,33 @@ public final class ProgramImageView extends BaseHtmlView {
     return publicStorageClient.getSignedUploadRequest(key, onSuccessRedirectUrl);
   }
 
-  private DivTag renderCurrentImage(ProgramDefinition program) {
-    if (program.summaryImageFileKey().isEmpty()) {
-      return div();
-    }
+  private DivTag renderCurrentProgramCard(Http.Request request, ProgramDefinition program) {
+    CiviFormProfile profile =
+        profileUtils
+            .currentUserProfile(request)
+            .orElseThrow(() -> new RuntimeException("Unable to resolve profile"));
+    Messages messages = messagesApi.preferred(request);
+    ApplicantService.ApplicantProgramData card =
+        ApplicantService.ApplicantProgramData.builder().setProgram(program).build();
+    LiTag programCard =
+        ProgramCardView.programCard(
+            request,
+            messages,
+            ApplicantPersonalInfo.ApplicantType.GUEST,
+            card,
+            Long.valueOf(profile.getId()),
+            messages.lang().toLocale(),
+            MessageKey.BUTTON_APPLY,
+            MessageKey.BUTTON_APPLY_SR,
+            /* nestedUnderSubheading= */ false,
+            layout.getBundle(request),
+            profile,
+            zoneId,
+            applicantRoutes,
+            profileUtils,
+            settingsManifest,
+            publicStorageClient);
 
-    String summaryImageFileKey = program.summaryImageFileKey().get();
-    if (!PublicFileNameFormatter.isFileKeyForPublicProgramImage(summaryImageFileKey)) {
-      return div();
-    }
-
-    ImgTag image =
-        img().withSrc(publicStorageClient.getPublicDisplayUrl(program.summaryImageFileKey().get()));
-    return div().with(image);
+    return div().with(h2("What the applicant will see").withClasses("mb-4")).with(programCard);
   }
 }
